requirements.txt
pyyaml
selenium
bs4
2captcha-python
colorama
python-telegram-bot